Agregar módulo de autenticación y configuración de rutas

+++ back end/customer-service/src/main/java/com/cibertec/dto/CustomerRequest.java
@@ -11,7 +11,7 @@ public record CustomerRequest(
         @NotNull(message = "userId no puede ser nulo")
         Long userId,
         @NotNull(message = "loyaltyPoints no puede ser nulo")
-        Long loyaltyPoints,
+        Integer loyaltyPoints,
         @NotNull(message = "birthDate no puede ser nulo")
         LocalDateTime birthDate,
         @NotBlank(message = "notes no puede estar vacio")
